refactoring: Add posts_dependency
